pyproject.toml
[project]
name = "career-bot"
version = "0.1.0"
description = "A professional career assistant chatbot powered by AI with multi-LLM support through unified OpenAI SDK"
authors = [{name = "Muhammad Ibrahim Malik"}]
readme = "README.md"
requires-python = ">=3.12"
keywords = ["chatbot", "ai", "career-assistant", "openai", "gemini", "gradio"]
dependencies = [
    "gradio>=5.0.0",
    "openai>=2.14.0",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
